# app/repository/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package db

import (
	"time"

	"github.com/google/uuid"
	"github.com/jackc/pgx/v5/pgtype"
)

type EmailVerificationToken struct {
	ID        uuid.UUID
	UserID    uuid.UUID
	TokenHash string
	ExpiresAt time.Time
	UsedAt    *time.Time
	CreatedAt time.Time
}

type OauthAccount struct {
	ID                uuid.UUID
	UserID            uuid.UUID
	Provider          string
	ProviderAccountID string
	Email             pgtype.Text
	CreatedAt         time.Time
	UpdatedAt         time.Time
}

type PasswordResetToken struct {
	ID        uuid.UUID
	UserID    uuid.UUID
	TokenHash string
	ExpiresAt time.Time
	UsedAt    *time.Time
	CreatedAt time.Time
}

type User struct {
	ID              uuid.UUID
	Email           string
	PasswordHash    *string
	FullName        string
	Role            string
	AvatarUrl       *string
	IsActive        bool
	EmailVerifiedAt *time.Time
	LastLoginAt     *time.Time
	CreatedAt       time.Time
	UpdatedAt       time.Time
}

type UserWord struct {
	UserID         uuid.UUID
	WordID         uuid.UUID
	Status         string
	Notes          *string
	IsFavourite    bool
	ReviewCount    int32
	CorrectCount   int32
	Box            int16
	LastReviewedAt *time.Time
	DueAt          time.Time
	AddedAt        time.Time
	UpdatedAt      time.Time
}

type Word struct {
	ID            uuid.UUID
	Term          string
	Language      string
	PartOfSpeech  *string
	Definition    string
	Example       *string
	Pronunciation *string
	Tags          []string
	CreatedBy     pgtype.UUID
	CreatedAt     time.Time
	UpdatedAt     time.Time
}