# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Backend API
  version: "1.0.0"
  description: |
    REST API for the Vite + React PWA.

    ## Response envelope

    Every response uses one of two shapes.

    Success:
    ```json
    { "success": true, "data": { }, "meta": { } }
    ```
    `meta` is present only on paginated collections.

    Error:
    ```json
    { "success": false, "message": "...", "code": "VALIDATION_ERROR", "errors": [], "request_id": "..." }
    ```
    `errors` carries per-field messages for validation failures. `request_id`
    matches the `X-Request-ID` response header and the server logs.

    ## Authentication

    Send the access token as `Authorization: Bearer <token>`. Access tokens are
    short lived (15 minutes by default); use `POST /auth/refresh` to rotate the
    refresh token into a new pair. Refresh tokens are single use: presenting one
    twice is treated as theft and revokes every session for the account.
  license:
    name: MIT
  contact:
    name: API support
    email: api@example.com
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
  - url: https://api.example.com/api/v1
    description: Production
...
security:
  - BearerAuth: []
paths:
  /auth/register:
    post:
      tags: [auth]
      summary: Register a new account
      description: |
        Creates a user with the `user` role and emails a verification link
        (logged, not actually sent, until a mailer is wired up). No session is
        issued: `POST /auth/login` is rejected until the address is verified.
        Rate limited to 10 requests per minute per IP.
      security: []
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
            example:
              email: jane@example.com
              password: Sup3rSecretPassword
              full_name: Jane Doe
      responses:
        "201":
          description: Account created, verification email sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserEnvelope"
              example:
                success: true
                data:
                  id: 3f0d1c1e-8f4e-4a4b-9c2a-2f9a1b7d9e11
                  email: jane@example.com
                  full_name: Jane Doe
                  role: user
                  avatar_url: null
                  is_active: true
                  email_verified_at: null
                  created_at: "2026-07-30T09:12:33Z"
                  updated_at: "2026-07-30T09:12:33Z"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /auth/login:
    post:
      tags: [auth]
      summary: Sign in
      description: |
        Exchanges credentials for a token pair. The response is identical for an
        unknown address and a wrong password, and both take the same time.
      security: []
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
            example:
              email: jane@example.com
              password: Sup3rSecretPassword
      responses:
        "200":
          description: Signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthEnvelope"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                success: false
                message: Invalid email or password
                code: UNAUTHORIZED
                request_id: 0f2b0e2c-3a1d-4c9a-9e21-6b7c8d9e0f11
        "403":
          description: The account is deactivated, or its email is not yet verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /auth/refresh:
    post:
      tags: [auth]
      summary: Rotate the session
      description: |
        Issues a new token pair and invalidates the presented refresh token.
        When `REFRESH_COOKIE_ENABLED` is on, the token may be omitted from the
        body and taken from the `refresh_token` cookie instead.
      security: []
      operationId: refresh
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
            example:
              refresh_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      responses:
        "200":
          description: New token pair
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthEnvelope"
        "401":
          description: Missing, expired or already-used refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                success: false
                message: Session expired, please sign in again
                code: UNAUTHORIZED
                request_id: 0f2b0e2c-3a1d-4c9a-9e21-6b7c8d9e0f11
  /auth/logout:
    post:
      tags: [auth]
      summary: Sign out of the current session
      description: Blacklists the presented access token and revokes the supplied refresh token.
      operationId: logout
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogoutRequest"
      responses:
        "200":
          description: Signed out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageEnvelope"
              example:
                success: true
                data:
                  message: Signed out
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/logout-all:
    post:
      tags: [auth]
      summary: Sign out everywhere
      description: Revokes every refresh token and invalidates outstanding access tokens for the account.
      operationId: logoutAll
      responses:
        "200":
          description: All sessions revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/change-password:
    post:
      tags: [auth]
      summary: Change the current password
      description: |
        Verifies the current password, replaces it, revokes every existing
        session and returns a fresh token pair so the caller stays signed in.
      operationId: changePassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
            example:
              current_password: Sup3rSecretPassword
              new_password: EvenB3tterPassword
      responses:
        "200":
          description: Password changed, new session issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthEnvelope"
        "401":
          description: The current password is wrong
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                success: false
                message: Current password is incorrect
                code: UNAUTHORIZED
                errors:
                  - field: current_password
                    message: Incorrect password
        "422":
          $ref: "#/components/responses/ValidationError"
  /auth/forgot-password:
    post:
      tags: [auth]
      summary: Request a password reset link
      description: |
        Always returns 200 with the same body, whether or not the address is
        registered, so the endpoint cannot be used to enumerate accounts.
      security: []
      operationId: forgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordRequest"
            example:
              email: jane@example.com
      responses:
        "200":
          description: Accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageEnvelope"
              example:
                success: true
                data:
                  message: If an account exists for that address, a reset link has been sent
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /auth/reset-password:
    post:
      tags: [auth]
      summary: Reset a password with a token
      description: Consumes a single-use token and revokes every existing session.
      security: []
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
            example:
              token: 7Qk2xR9vN0pLm3sT8wYzA1bC4dE5fG6hJ7kL8mN9oP0
              new_password: EvenB3tterPassword
      responses:
        "200":
          description: Password updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageEnvelope"
        "400":
          description: The link is invalid, expired or already used
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                success: false
                message: This reset link is invalid or has expired
                code: BAD_REQUEST
  /auth/verify-email:
    post:
      tags: [auth]
      summary: Verify an email address with a token
      description: Consumes a single-use token sent at registration. Required before login.
      security: []
      operationId: verifyEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyEmailRequest"
            example:
              token: 7Qk2xR9vN0pLm3sT8wYzA1bC4dE5fG6hJ7kL8mN9oP0
      responses:
        "200":
          description: Email verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageEnvelope"
              example:
                success: true
                data:
                  message: Email verified, you can now sign in
        "400":
          description: The link is invalid, expired or already used
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                success: false
                message: This verification link is invalid or has expired
                code: BAD_REQUEST
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Access token from /auth/login. Header: `Authorization: Bearer <token>`."
  schemas:
    User:
      type: object
      required: [id, email, full_name, role, is_active, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        full_name: { type: string }
        role: { type: string, enum: [user, admin] }
        avatar_url: { type: string, nullable: true }
        is_active: { type: boolean }
        email_verified_at: { type: string, format: date-time, nullable: true }
        last_login_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    Tokens:
      type: object
      properties:
        access_token: { type: string }
        refresh_token: { type: string }
        token_type: { type: string, example: Bearer }
        expires_in:
          type: integer
          description: Access token lifetime in seconds.
          example: 900
        expires_at: { type: string, format: date-time }
...
    FieldError:
      type: object
      properties:
        field: { type: string, example: email }
        message: { type: string, example: Must be a valid email address }
    ErrorEnvelope:
      type: object
      required: [success, message]
      properties:
        success: { type: boolean, example: false }
        message: { type: string }
        code:
          type: string
          enum:
            - BAD_REQUEST
            - VALIDATION_ERROR
            - UNAUTHORIZED
            - FORBIDDEN
            - NOT_FOUND
            - CONFLICT
            - TOO_MANY_REQUESTS
            - PAYLOAD_TOO_LARGE
            - UNSUPPORTED_MEDIA_TYPE
            - INTERNAL_ERROR
            - SERVICE_UNAVAILABLE
        errors:
          type: array
          items:
            $ref: "#/components/schemas/FieldError"
        request_id: { type: string }
    MessageEnvelope:
      type: object
      properties:
        success: { type: boolean, example: true }
        data:
          type: object
          properties:
            message: { type: string }
    AuthEnvelope:
      type: object
      properties:
        success: { type: boolean, example: true }
        data:
          type: object
          properties:
            user:
              $ref: "#/components/schemas/User"
            tokens:
              $ref: "#/components/schemas/Tokens"
    UserEnvelope:
      type: object
      properties:
        success: { type: boolean, example: true }
        data:
          $ref: "#/components/schemas/User"
...
    RegisterRequest:
      type: object
      required: [email, password, full_name]
      properties:
        email: { type: string, format: email, maxLength: 254 }
        password:
          type: string
          minLength: 12
          maxLength: 72
          description: Must contain an uppercase letter, a lowercase letter and a digit.
        full_name: { type: string, minLength: 2, maxLength: 120 }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, maxLength: 72 }
    RefreshRequest:
      type: object
      properties:
        refresh_token:
          type: string
          description: Optional when the refresh cookie is enabled.
    LogoutRequest:
      type: object
      properties:
        refresh_token: { type: string }
    ChangePasswordRequest:
      type: object
      required: [current_password, new_password]
      properties:
        current_password: { type: string, maxLength: 72 }
        new_password: { type: string, minLength: 12, maxLength: 72 }
    ForgotPasswordRequest:
      type: object
      required: [email]
      properties:
        email: { type: string, format: email }
    ResetPasswordRequest:
      type: object
      required: [token, new_password]
      properties:
        token: { type: string, minLength: 16, maxLength: 256 }
        new_password: { type: string, minLength: 12, maxLength: 72 }
    VerifyEmailRequest:
      type: object
      required: [token]
      properties:
        token: { type: string, minLength: 16, maxLength: 256 }
...
  responses:
    Unauthorized:
      description: Authentication is required or the token is no longer valid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            success: false
            message: Authentication is required
            code: UNAUTHORIZED
...
    Conflict:
      description: The resource already exists
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            success: false
            message: An account with this email already exists
            code: CONFLICT
    ValidationError:
      description: One or more fields are invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            success: false
            message: The request contains invalid fields
            code: VALIDATION_ERROR
            errors:
              - field: email
                message: Must be a valid email address
              - field: password
                message: Must be at least 12 characters
            request_id: 0f2b0e2c-3a1d-4c9a-9e21-6b7c8d9e0f11
    TooManyRequests:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema: { type: integer }
          description: Seconds until the window resets.
        X-RateLimit-Limit:
          schema: { type: integer }
        X-RateLimit-Remaining:
          schema: { type: integer }
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            success: false
            message: Too many requests, please slow down
            code: TOO_MANY_REQUESTS